+++ resources/lang/fr/validation.php
@@ -1,0 +1,32 @@
+<?php
+
+/**
+ * @package laragrad/identifier-validation
+ * @desc EN language file
+ */
+return [
+
+    // RUS
+    'rus_inn' => 'The :attribute is not a valid company INN',
+    'rus_person_inn' => 'The :attribute is not a valid person INN',
+    'rus_kpp' => 'The :attribute is not a valid company KPP',
+    'rus_ogrn' => 'The :attribute is not a valid OGRN',
+    'rus_ogrnip' => 'The :attribute is not a valid OGRNIP',
+    'rus_snils' => 'The :attribute is not a valid SNILS',
+
+    // KAZ
+    'kaz_iin' => 'The :attribute is not a valid IIN',
+    'kaz_bin' => 'The :attribute is not a valid BIN',
+
+    // BLR
+    'blr_unp' => 'The :attribute is not a valid UNP',
+    'blr_person_unp' => 'The :attribute is not a valid person UNP',
+
+    // UKR
+    'ukr_edrpou' => 'The :attribute is not a valid EDRPOU',
+
+    // Other
+    'bank_card_number' => 'The :attribute is not a valid bank card number',
+    'isin' => 'The :attribute is not a valid ISIN',
+    'iban' => 'The :attribute is not a valid IBAN',
+];